Move advertiser.edit to advertiser.edit.index
