+++ app/gql/queries/advertiser/edit.graphql
@@ -1,7 +1,7 @@
-#import 'fortnight/gql/fragments/advertiser/edit'
-
 query EditAdvertiser($input: ModelIdInput!) {
   advertiser(input: $input) {
-    ...AdvertiserEditFragment
+    id
+    name
+    deleted
   }
 }
