+++ app/templates/manage/advertiser/edit/index.hbs
@@ -1,0 +1,34 @@
+{{#model-form
+  model=model
+  shouldAutosave=true
+  onSubmit=(action "update") as |form|
+}}
+
+  <div class="card-body">
+    {{ partial "manage/advertiser/form" }}
+  </div>
+  <div class="card-footer">
+    {{#link-to "portal" model.hash class="btn btn-info"}}{{entypo-icon "eye"}} Portal{{/link-to}}
+
+    <div class="btn-group float-right">
+      {{#if model.deleted}}
+        {{confirm-button
+          icon="cw"
+          label="Restore (Undelete)"
+          class="btn-success"
+          disabled=isActionRunning
+          onConfirm=(action "undelete")
+        }}
+      {{else}}
+        {{confirm-button
+          icon="trash"
+          label="Delete"
+          class="btn-danger"
+          disabled=isActionRunning
+          onConfirm=(action "delete")
+        }}
+      {{/if}}
+    </div>
+  </div>
+
+{{/model-form}}
